career-counsellor_using_livekit_ADK_RAG/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==15.1.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-genai==1.47.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
livekit==1.0.17
livekit-agents==1.2.17
livekit-api==1.0.7
livekit-blingfire==1.0.0
livekit-plugins-deepgram==1.2.17
livekit-plugins-google==1.2.17
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-speechify==1.2.17
livekit-protocol==1.0.8
MarkupSafe==3.0.3
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
pillow==11.3.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sounddevice==0.5.3
tenacity==9.1.2
tqdm==4.67.1
types-protobuf==6.32.1.20250918
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
watchfiles==1.1.1
websockets==14.2
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
google-search-results==2.4.2
langchain-community==0.4.1
google-adk
faiss-cpu==1.10.0 
google-generativeai==0.8.4